+++ app/src/main/kotlin/me/xizzhu/android/joshua/settings/SettingsActivity.kt
@@ -39,9 +39,9 @@ class SettingsActivity : BaseActivity() {
                 SettingsViewHolder(
                         findViewById(R.id.display), findViewById(R.id.font_size), findViewById(R.id.keep_screen_on),
                         findViewById(R.id.night_mode_on), findViewById(R.id.reading), findViewById(R.id.simple_reading_mode),
-                        findViewById(R.id.hide_search_button), findViewById(R.id.consolidated_sharing),
-                        findViewById(R.id.default_highlight_color), findViewById(R.id.backup_restore), findViewById(R.id.backup),
-                        findViewById(R.id.restore), findViewById(R.id.about), findViewById(R.id.rate), findViewById(R.id.version)
+                        findViewById(R.id.hide_search_button), findViewById(R.id.consolidated_sharing), findViewById(R.id.default_highlight_color),
+                        findViewById(R.id.backup_restore), findViewById(R.id.backup), findViewById(R.id.restore),
+                        findViewById(R.id.about), findViewById(R.id.rate), findViewById(R.id.website), findViewById(R.id.version)
                 )
         )
     }